noop merge scroll and touch handlers
